backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Cache / Queue
redis==5.1.1
celery[redis]==5.4.0

# Data fetching
httpx==0.27.2
aiohttp==3.10.10
yfinance==0.2.44
fredapi==0.5.2
wbgapi==1.0.12          # World Bank API
pandas==2.2.3
numpy==2.1.2
polars==1.9.0

# Phase 2 additions
websockets==13.1         # Binance WS consumer
beautifulsoup4==4.12.3   # Investing.com calendar scraper
lxml==5.3.0              # Fast HTML parser for BS4
openpyxl==3.1.5          # AAII Excel file parsing

# Utils
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0
pytz==2024.2
tenacity==9.0.0         # Retry logic
structlog==24.4.0       # Structured logging
